refactor: Do not fail E2E tests if no configs can be found for cleanup
We already did not fail teste but only log if a single test configuration can not be cleaned-up,
while still failing the whole test if the GET request to list all configs fails.
To make this more lenient and reduce false positive test failures furhter, a failure
to find configs of a given API for cleanup is now also just logged.