feat(convert): Inform users when conversion will result in new API handling even though config type name does not change
Several new config -v2 types where introduced in 1.x to not make the new non-unique-name handling default behaviour, as it
required manual interaction.

As this handling becomes default with 2.x, but the API name does not change there was no logging so far, informating the
user of the 'conversion' resulting in different handling than before.

Add logging for this case including a link to documentation.
To be safe it's checked that the deprecating API is actually a non-unique name one, but they all are.