refactor: Move upsert non-unique-name logic from deploy into client
Our current level of abstraction puts all logic around how configurations are created on the
Dynatrace API into the rest pkg.

Only the non-unique-name log was currently in the deploy pkg.
This moves the existing code without any logical changes into the rest pkg
as part of the config_client.