refactor(rest): Split low-level REST API and higher level Client logic into separate package
Previously the rest package contained everything that dealt with communication to Dynatrace,
from actual HTTP requests and understanding details like pagination headers,
to the actual monaco logic of how configurations are created or updated.

This change splits the package into a 'rest' package - which retains the low-level Dynatrace REST API
handling logic, and a 'client' package that contains the 'Dynatrace Client' logic of how to upsert
Configs or Settings.