feat(client): Add 404 to list of known retry errors for applications
Immediately after extending the handling to include 409 errors, E2E tests ran into a case
where a Web application update - meaning the meID was found by GET all - fails with a 404
not found. It may be that the application API was actually fixed to return proper error codes
for cases it previously returned 5xx errors handled by the retry logic so far.