fix: windows can't store files with ':' in the name
Settings schemas have the structure <prefix>:<name> where the colon splits the two parts.
During download, we store all downloaded configs in folder names with those exact names. On Linux, this is no issue, but Windows does forbid colons, as they are reserved for drive names.

To mitigate this, we now sanitize the naming during persisting.