ci(e2e): Actually match checkout conditions to run e2e build step
Using the simple condition that is applied to e2e tests is not enough to ensure that
binaries are only build if code was checked out/the build was on main, scheduled or labeled
to run a PR e2e test.

To ensure this works, the release binary step is run conditionally when either the
base repo checkout condition, or the PR checkout conditions are met and there is
actually code to build.