perf: Use Aho–Corasick
Instead of searching for references using strings.Contains, we now use the Aho–Corasick-algorithm to search for references efficiently.
The algorithm has the advantage that it is efficient by first building a state-machine using the ids as dictionary and returning all occurrences within a search string.

To add this behavior, a new dependency is introduced that implements this behavior