feat(e2e): Include Settings in nightly test cleanup
Add simple Settings cleanup removing all Settings with externalID.
While this a bit more than the Config cleanup that is limited to
those matching the format of E2E test created configs, the test
envs that this cleanup is used against, should not have any Settings
created by monaco that did not come from a test.

For an exact cleanup we would need to know configIDs to then generate
the correct externalID, which not possible with the random suffix of
test configurations.