test(e2e): Make test suffix available as env var
Expose the unique test suffix as an envrionment variable so it can be used
in directly in test configuration properties that need to be made unique,
but are not rewritten but the default rewrite of name and config ID.

Additionally the suffix is now lower-cased by default as some usages of
it via env var might not allow upper case letters. For the sake of
general test uniqueness the case of the suffix does not matter.