feat(client): More leniently handle pagination failures
Testing has shown that it is possible to fail paginated GET requests if something
like another E2E test, or a user, removes enough configurations to remove a page,
after a call has received a nextPageKey. This will result in querying for a no
longer existing page and result in a 400 error reply.

As in these cases we did in fact fetch all the data that is available, 400 errors
on requests with a nextPageKey are no longer handled as an error, but instead as
a warning and sign to stop querying for more.
There is a very limited chance that this will swallow actual invalid requests, but
for that the API would need to change making our current query invalid.