test(e2e): Assert created configs instead of logs in pagination test
The nightly pagination test didn't work anymore since we changed the Settings
upsert mechanism and can no longer discern between updates and creation and the
test asserts were based on log output alone.
Instead the test now asserts the number of logged API calls for create/update
as well as that all test objects are available on the environment.