refactor(client): Remove client from environment defition util
The utility method to create a Dynatrace client based on an environment definition directly,
introduced an otherwise unecessary dependency from the client to the manifest.

As the client is not instantiated in many places, the function is removed and clients
are instantiated with data read from the environment definition manually again.