refactor: Introduce UrlDefinition.Name
Currently UrlDefinition.Value has two purposes. In case UrlDefinition.Type is 'value', it's the actual value of the url, otherwise it's the environment-variable-name of the url that has to be resolved later.

Now we have three fields. The 'Type' as is, 'Value' as the resolved value (for env vars it is resolved during manifest loading), and 'Name' which is the Env-var-name. We need it for conversion.

Additionally, we can now remove the logic from environment.GetUrl to only return the resolved value.