refactor: Move NewEnvironmentDefinitionFromV1 to convert package
The method NewEnvironmentDefinitionFromV1 builds a new v2-environment for a given v1-environment.
It does not make sense that the method (and interface) is defined in the manifest package, but now is moved to the convert-package. With that, the environmentv1-interface is also no longer needed and the method can be private.

Additionally, this removes all v1 dependencies from the manifest.go file