feat(delete): Remove restriction on delete yaml filename
The is no clear reason why 2.x should retain the restriction of delete yamls,
being exactly named delete.yaml. Unlike 1.x the delete command accepts any file
path and will load and parse delete entries from it, instead of looking for a
defined 'delete.yaml' in the folder 'monaco deploy' is being executed in.

In addition:
- the filename restriction did not validate the file, but the full path,
meaning delete had to be executed in the folder containing the delete.yaml
for input validation not to fail
- the delete loaded needlessly split and recombined a given filepath into a
directory and the filename. This is replaced by simply trying to get the absoulte
path to the given delete yaml file.