fix(config): Add 'name' to list of reserved parameters
Properties of the toplevel Config object must not be re-defined as parameters,
but 'name' was missing from the slice that is used to validate this.

As conversion also uses this slice in two places to ensure conversion does not
produce invalid overlaps with v2 reserved parameter names, the exclusion
for 'name' thas was already in place at one callsite is moved to the method
called in both cases.
Additionally the method is documented and renamed to make it clear(er) what
it is meant to do.