refactor: Move loading of default TokenURL to client
We decided that the responsibility of loading the default Dynatrace SSO
token URL should be encapsulated in the Dynatrace Client, rather than
manifest loading.
This is achieved by having the manifest treat the token URL as a pointer,
and loading the default URL when nothing was given to the client on
instantiation.