refactor(download): Use Client interface in download client provider rather than implementation
As Downloading does not actually need the DynatraceClient implemenation and to allow using Client mocks in tests,
the download options' DynatraceClientProvider function is changed to return the Client interface, rather than
a pointer to the concrete DynatraceClient implementation.

To keep usage simple, a default provider function is defined, that just wraps the call to NewDynatraceClient and
abstracts to the interface.