feat(manifest): Trim trailing slashes from URLs when loading manifests
As several parts of the Client code append API paths to the environment URL directly,
the risk of producing invalid URLs by appending a slash to an URL that already ends
with a slash, is reduced by trimming slashes on manifest loading.

In the long run all client code should be made resilient to this in a similar way
that metadata and version checks are, by using url.URL path joining which covers
this case correctly and does not result in double slashes, even if both host and
path end/start with a slash.