fix: Use config name for created Settings objects
As Settings do not have a general 'name' field, the client currently
just returns the objectID as both name and ID of created entities.

This causes problems if a user tries to reference the name of Settings
configuration in another config.
As we still require every configuration to define a name, this change
simply extracts the name from the configuration rather than using data
returned from the API/client, and sets it when deploying.

To not break if we ever decide the make the name property optional for
Settings, a fallback to using the ID - but doing so explicitly with a
warning - is implemented.