feat: Make config.Config.Type an interface
Until now, config.Type has been a struct that contains all possible fields. To check what type the config has, several isX methods were implemented. This is convenient for a few types, but with more and more unmaintainable.

This change makes config.Type an interface that all types can implement.
This allows us to add more types and only change a few places where it is necessary to distinguish the types.
One decision was to force the users of the method 'deploy' to cast and check the type independently. One alternative would have been to add generics to the config object. To accomplish this, we would have to change all methods where the config.Config type is used, and even then, we're unsure if this would work.
Another alternative would have been to add a method, 'getData' to the Type interface. But this accomplishes basically the same, as the return value of this function would have to be checked and parsed too.