fix(download): Limit manifest loading to specified environments
Like other commands download needs to pass the specific environment it is running
against to the manifest loader, to ensure that unused environments don't get parsed
and require environment variables to be set for their tokens (and possibly URLs) which
won't be used.

It appears download was missed, when limited loading was introduced to cope with the
fact that Environment variable resolution was moved from 'at use' to 'at load' time.