fix(download): Use Coordinate ID when resolving dependencies
The previously used Template ID just happened to be equal to the config's ID when downloading,
if the name of downloaded templates ever changes the resolution logic would break.
The config ID is a more reasonable source for the ID, even though 'downloaded configs use the
object ID as config ID' is also just convention (and not the case for Settings).