fix(download): Resolve Managment Zone references by numeric ID
Management zones are always referenced in configurations via their numeric
ID instead of a Config UUID or Settings objectID for backwards compatibility
reasons.

To still allow download code to correctly build reference parameters to
management zone Settings configurations, which are only known by their objectID,
code is introduced that leverages knowledge about how the numeric ID is formed
internally and transforms a given object ID into a numeric ID.

The implementation matches the logic of the Dynatrace implementation for
this as closely as possible.