build: Do not include 'netgo' build tag in windows release-builds
If the netgo build-tag is given, the pure Go resolver is forced to be used instead of the cgo provider.
While this works fine most of the time, complex DNS configurations are not yet supported by the Go resolver.
Since this issue only exists for windows, and the netgo tag is required for minimal docker-images (e.g. alpine), we now skip the tag for windows, but keep it for linux builds.