build: Build the correct binary
Up to now, the output binary was always `monaco-windows-amd64.exe`, no matter what the actual binary was. This was no problem as our build-pipeline builds one binary at a time, but for local testing this is not useful.
This commit fixes the output file so that each binary is correctly created.