ci(dep-and-licenses): Use plain assets URL from release event for asset upload
Rather than the upload_url - which is a hypermedia resource template and thus not directly usable
as our cUrl call assumed - use the plain assets URL also provided in the event.

This is less robust to changes of the API than actually parsing the resource template, but functional
and less effort.

Note that while Github Actions exist to upload assets to releases, all of them are 3rd party actions -
thus it was decided to keep this as a simple cUrl call rather than adding a new dependency to an action
that's not directly supported/updated by GitHub either.