feat(download): Allow switching between fast and basic dependency resolvers
It was found that the currently used 'fast' dependency resolution is using excessive
memory.

As a quick fix for customers running into OOM errors in some cases, this change makes it
possible to switch back to the old naive implementation which is more CPU intensive and thus
slower, but has a much lower memory footprint.

Control of wich implementation is used is done via a new MONACO_FEAT_USE_FAST_DEPENDENCY_RESOLUTION
feature flag.

This is somewhat of a band-aid fix, until a resolution algorithm/implemementation is found that
combines acceptable memory and CPU usage.