ci: Build release asset upload URL directly for BOM upload action
The previously used release.asset_url was also wrong for uploads,
as it just contains the api.github URL at which assets can be downloaded.

Rather than parsing the Hypermedia URL that's contained in upload_url,
this simply builds the URL directly using just the release ID from the
event.

While this is not resilient to this URL changing, we do the same in our
general release build - which has no access to github event data - so if
this breaks, manual intervention to our release pipelines is needed anyway.