feat(deploy): Log number of projects, environments and (on debug) configs to deploy
It is sometimes helpful to get a quick overview of the number of projects and configurations
that are being deployed by just looking at the logs.

As collecting the number of configs requires counting them by running through all configs
per project, environment and type, this is limited to only happen when debug logging is
active.