ci: Publish containers on docker-hub
This change introduces releasing containers on docker-hub, and also setting the latest-tag.

First of all, we try to find what kind of release this is, either full-, rc-, or dev-release. If it is neither, we fail the build to restrict the possible tags.

Then we release monaco to docker for full- and rc-releases. Additionally, we publish the latest tag for those releases.

Furthermore, we push the latest-tag also for our internal releases, but only if it's not a dev-release.