feat(download): Sanitize downloaded Automation configurations
Like existing types, Automation payloads contain data that is not needed on upload
and just clutters config-as-code files - thus a simple sanitization is added.
As these APIs are fewer and more consisten than Classic Config APIs, the full fledged
sanitizer code from their download is not reused.

Additionally the 'title' field which all currently supported Automation types contain,
is extracted as the config's 'name' if it is found.

To make this resilient, any failure in extraction is just treated as before and will
persist the payload as is.