chore: Replace GitHub issues with links to Dynatrace resources
Remove the feature-request template and modify the bug-report template to ask for bugs
that the reporter already can describe in detail.

Add new links to the create issue template chooser which link to support and community resources.

To better enforce our contributing guide's ask for creating a discussion before
creating a PR, a link to discussions is added as well.

For details on the contact_links configuration see: https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository#configuring-the-template-chooser