refactor(json): Remove unused error return from MarshalIndent utility
As the method already falls back to returning unchanged input if indenting it failed for any reason,
and the error return was ignored in all production code usages, it is simply dropped.

Functionally monaco will still behave the same, printing a warning and writing unindented files in
case a downloaded JSON could not be indented as expected.