ci: Split Docker registry and repo path into separate inputs
To ensure we log in to the registry correctly (login via sub-path, seems to cause issues
for Docker Hub), the registry host and repo/project name are split into dedicated values,
rather than having a single full path as input.

This should ensure that logins work, regardless of which regstriy we target, and that
projects/repos can change independetly by modifying a single configuration.