fix(rest): Re-introduce custom user-agent at the lowest REST lib layer
With introducing oAuth support, we have lost the dedicated user-agent from the
oAuth client (never added) as well as API token authenticated connects (not sent/
overwritten somewhere after the Transport struct it was added to).

To ensure we don't forget to send this with any future client implementations,
setting the User-Agent header is moved to the lowest possible layer in our API
call methods, the final rest/request method that actually calls the http/client.