feat(delete): Introduce structured delete definition
As the 'externalID' by which Settings are identified by monaco is extended
to include the project (full coordinate - project,type,id) - the simple
string format is no longer enough to define delete entry.

A new format is added that defines a structured project, type and name or ID
delete entry.
Loading is extended to be able to load both the full and the old simple format.