feat(convert): Include the api in the duplicate lookup
This change includes the api-id in the duplicate lookup. This means that the id is not increased if the id is duplicated over multiple APIs.

Another difference this change brings is that the first config will keep the original id, while the first duplicated config will have the new id.