feat: Write downloaded configs in a stable order
Monaco is often used to download configurations from the same environment at two different points in time. Until now, it was pretty hard to compare the configs as they were written in the order we downloaded them.

This change orders the configs before writing them to files, thus simplifying comparing incremental changes or the difference between two environments.
We achieve the ordering by simply ordering by the config-id within a file, as the API already splits downloaded files, and thus, the ID is unique within a file.