refactor(automation): Simplify import names of automation client
The client was imported as 'automationClient' in download/automation and
that import name was reused in other places, adding a useless 'Client'  to
the name.

Where possible the automation client is imported without alias as 'automation'.
In download/automation the import alias is simplified to 'client' as the extra
'automation' does not add valuable context.