refactor(client): Introduce a single ClientSet to remove instantiation logic out of specific cmds or pkgs
Previously commands and packages shared some logic on instantiation of the base DynatraceClient used to access
Classic Config and Settings APIs, but all implemented specicif setups of the new AutomationClient.

This takes the ClientSet concept that was introduced for the deploy package and moves it to the client/ package.
Additionally it unifies the client setup - in some places we added optional version checks, while in others we did
not. This also means that the concurrency limiter so far only applied to downloads is now on all clients - which might
be helpful/ensure we can't forget about it when adding deploy concurrency.

Wherever an API client is used ('function' packages) a dedicated ClientSet struct is introduced and prepared for a
possible followup to introduce dedicated interfaces per consumer.
For now all pkgs retain usage of the DynatraceClient 'fat interface' and only introduce a dedicated AutomationClient
interface for theit usecases.