feat(download): Skip unmodifiable settings objects
Some Settings 2.0 objects can not be modified - e.g. Dynatrace default configurations.
When downloading these, the resulting project can not be deployed again successfully
as trying to change these unmodifable objects will fail.

As the Settings API now reports a modificationInfo object if requested, the Settings
client is extended to query and return this info.
Older Dynatrace versions may not return this field yet, thus it is handled as a pointer
to clearly indicate that it may not be present on a downloaded settings object.

The Settings downloader is extended to filter out unmodifiable Settings instead of
persisting them.