refactor(download): Change warnings about discarded settings to debug
The fact that a Settings object was filtered out is not actually a problem but intended behaviour.

To match the log level of filtering Classic Configs it is reduced to Debug - for a standard user
downloading configuration, the fact that some things were filtered out do not provide value.

Additionally the log is simplified.