fix: Allow backslashes in template paths
When a project is created by a Windows user, the paths might contains backslashes (`\`) instead of slashes (`/`) for path-separation. This is not an issue if the project is deployed in Windows again, but if it's deployed on a Linux environment, for example in Linux-CI systems, the path could not be resolved.

This bugfix allows backslashes in template paths, as a Windows-user might also manually create the path, and Monaco should handle it correctly.