feat: introduce logger interface and console and zap implementation
the "old" log implementation is replaces with on backed by zap.
I am using zap to perform structured logging to file and console.

The request / response logging was moved into its dedicated package,
as it cannot be seen as "regular logging".

further, timestamps are now matching with what is shown in the logs and how the download folder and log file are titled.