ci: Fix tagging container releases as 'latest'
Previously we accidentally built another binary and container with version
set to 'latest' instead of adding a tag to the image for the specific release
version.

This is replaced by a conditional tagging step in createContainerAndPushToStorage,
which will apply the 'latest' tag after building and pushing the image if configured.