feat(config): Fail early if a Classic Config API config references anything but another Config API type
Config API type configurations generally have a UUID or Monitored Entity ID, while newer configuration types like
Settings use IDs that are not compatible with what the Config API expects.

As users are migrating existing configuration over to Settings it happens that an existing Config API configuration
references the ID of something now managed as a Setting.
Deploying such a configuration will fail withe API reporting that the Settings objectID is not a valid ID - Config APIs
are not able to resolve the new base64 encoded objectID, while the Settings API accepts any type of ID.

To warn users as early as possible, such invalid references result in an error as soon as the configuration is loaded.

To be lenient, two exceptions exist:
- validation only applies to referencing a config's ID, any other properties can be referenced regardless of type.
- referencing a management-zone Settings's ID is allowed, as they get converted to numeric IDs which are shared between Config and Settings API