build: Compile sources as well as tests in 'compile' make target
Previously compile only touched sources, as go build ignores any test files.
To ensure the make target actually works as it was intended a second step is
added to compile tests as well.

While go test offers a -c flag which limits it to compiling, rather than executing
tags, this only works for single tests or packages, not for several.

To still be able to run go test ./... to compile ALL tests, a workaround is applied,
which 'runs' a specific test that does not exist. This ensures go test will compile all
tests in a package while trying to find the specific tests - which it won't.

see: https://stackoverflow.com/a/72722257