ci: Sign image tag rather than repo digest
Querying the repo digest(s) and using the first entry in the images list of
repo digests, we accidentally always sign and deliver on the first digest the
release image was built for. For a full release that is the Docker Hub release,
which will receive a signature twice, while the internally delivered image will
not receive a signature.

As there is no recommendation to sign based on the full sha rather than a tag that
could be found in cosign documentation, this issue is resolved by simply signing the
tag.
As we already use the tag to query for the full digest sha-sum, there is also no benefit
to authenticity by using the checksum - it's already that of the tag (which we can strongly
assume was just built and pushed by us anyway).