fix: Consider filtering for counting entries received paginated responses
Previously versions of the 'addToResults' method used when querying paginated APIs returned
the number of elements that were read from the API payload and added to the full results list.

This does not work correctly for implementations that include filtering logic, as the client
might receive the expected number of elements, but only include a smaller subset in the result
slice after filtering.

This results in some usages - like when querying for a specific Setting which filters on externalID -
printing incorrect warning logs about missing entries.