feat(download): Extract monitored entity IDs into YAML parameters
To make it easier for users to transfer configurations and modify environment specific values,
like monitored entity IDs, these are now extracted from the downloaded JSON payloads and stored
as YAML parameters.

This allows a user - or automated system - to work with the YAML files alone when modifying ME
IDs.