fix(log): Ensure structured logs only contain fields from context that actually exist
Previously if a context was missing either environment or coordinate data (this is likely),
the structured logs would still contain null fields.
Changing the context fields slice from a fixed size to a fixed capacity this problem is solved
and only fields actually present will be part of the structured logs.