refactor(client): Set User-Agent string via Transport RoundTripper rather than explictly per request
A new CustomUserAgentTransport is introduced and used to wrap the Classic and OAuth client's transports
if a custom user agent is to be sent.

For usage convenience the ClientSet is extended with options to set a user-agent on this level, but in
current monaco CLI production use will default to the existing 'Dynatrace Configuration as Code' UA string.

To ensure that request traffic logging includes the user-agent that's actually sent, the request is logged
after it was sent (and the roundtrippers have modified headers), rather than before.
To still be able to log request bodies, these are not extracted manually before the request is executed as
sending drains and closes the request body.