refactor: Ensure custom errors use uniform names and have struct tags for JSON logging
To ensure that structured JSON logs are easy to parse, fields of custom errors that have the same
meaning now have the same name.

For a nicer looking generated JSON all fields are extended with struct tags so they start lowercase,
rather than than the upper-case of exported struct fields.