test: Allow overwriting the workflow-actor in e2e tests
The workflow actor can be different for different tenants/accounts.
While a customer can use environment-overwrites to set the correct value, we currently re-use the testing environments for different stages, thus unable to use the environment-overwrite.
To simplify this code, we'd need to use some kind of context that we can overwrite for each environment.