ci: Update manifests to allow overwriting the token endpoint
The sso-token endpoint is always the same for production, so we can assume it as a default value. For hardening tests though we need to overwrite it using environment-variables. This also means, that we now need to set the token endpoint in prod-e2e tests, thus setting the env variable in the new Jenkins build, but also as a GitHub secret.