feat: Implement function to deploy configs using the graph
Instead of using only the topological sort for all configs, the new method uses the independently components and deploys the configs based on them.
The algorithm works as follows:
All independent components are deployed independently. For example, no information about resolved parameters is passed between the components.
Within each component, the configs are deployed based on the toplologically sorted nodes. In case a config fails to be deployed or is skipped, all children of this config are marked as 'can't deploy'. Configs, which can still be deployed within the compnent will still be deployed.
Based on the continue-on-error flag, the component might exit early.