feat: Add command allowing to generate a delete file from a given manifest
A new 'monaco generate deletefile' command makes it easy to create a delete file for a
given project.
Generally it's assumed that this file will be a starting point for creating an actual file,
but it could directly be used to remove everything a manifest created.

In this simple form of the command names that are a reference to other configs are not resolved.