feat(manifest): Allow loading manifests without resolving environment variables
For some usecases we don not actually need a fully valid manifest, but just care about it's projects,
for this a new LoaderOption is introduced, which allows skipping the resolution of environment variables
that may be used in URLs and Secrets.

By default they will be loaded and no existing code paths need to be adapted.
New codepaths can however explicitly deactive resolution by setting an option boolean to true.