chore: Ensure all-configs SLOs are still valid if test-suffix is empty
All configs test resources are a useful sample to check things against a test environment
as they define all types we support and use most monaco features.

To be able to deploy them without the name-rewriting that happens on E2E tests, a 'UNIQUE_TEST_SUFFIX'
env var needs to be set to empty string (used in some references so setting an actual value fails if no
re-write happened).

Setting this to an empty string would produce invalid SLO configurations at they end in an underscore,
which the API does not allow.
As it does not matter for the test itself, the underscore separator is simply removed, ensuring the
local usage of the project works.