fix(delete): Ensure users are only warned about missing oauth if platform types are defined
Also the logic that automation deletion isn't attempted if no client is defined is fixed
(the interface is not nil, but the underlying pointer may be. Previous code panics)

While logic could be changed to return an error if platform types are in the file but
the client from the manifest is for a classic env, delete has always been lenient in
what it deletes and users might define long delete files that remove things from disparate
environments. It seems more user friendly to also keep the fact that no automations will
be deleted as a warning.