chore(http): Improve API token format warning
The previous log about using an old token format was misleading, as we don't
actually check for an old format, but just if the new format is matched.

As the new format exists since October 2020, it is much more likely that a
random string was provided as a token, than an actual old API token.

To reflect this, the log warning is changed to point the user towards checking
their configuration first, but still mention to possibly update their tokens.