fix: Ensure support-archive is written regardless of command success/error
Due to a known limitiation of cobra, we did not create support-archives in case a command
failed.
PersistentPostRun methods of a command are not executed if running the command produced and error.
See: https://github.com/spf13/cobra/issues/1893

To ensure a support-archive is always created if requested, the method is instead registered as
a global cobra finalizer and will run after any cobra command finished, regardless of status.