refactor(deploy): Reimplement simple per-config name uniqueness check for classic cfgs in sequential deploy
As the name uniqueness check was removed from classic.Deploy but is still possible to do for old sequential deployments,
it is re-implemented in sequential.DeployConfigs.
While this does duplicate a bit of setup code, the sequential package will only live until graph based deployments
become fully default and the feature flag to go back is removed.

Retaining the old checks for the old style of deployments keeps it fully independet of any of the changes introduced for
graph based (parallel) deployments, ensuring we can fully fall back to old behaviour if needed.