fix(trafficlogs): Create log folder if none exists before trying to write traffic logs
While it usually does not happen, as the base logger setup will create the .logs/ folder, it is
still safer to check and possibly create the folder before trying to write traffic log files.

Additionally in tests it can happen that no .logs/ folder exists, as traffic logs always write to
the OsFs while the base logger may be writing somewhere else (MemMap fs or temp file fs with a
different path).