refactor: Simplify resolving parameter values by no longer requiring a sorted list of parameters
As resolving parameter values of a configuration always requires its parameters to be sorted,
so that possible references between parameters can be resolved, this sorting is now done in
resolve.ParameterValues itself instead of requiring callers to provide a pre-sorted list of the
configuration's parameters.
This also removes the edge-case risk that a caller accidentally supplies a list of sorted parameters
that are not actually from the given configuration.

This change allows further cleanup of pkg/deploy's internal subpackages, removing it's own wrapper
function sorting and resolving parameter values, and subsequently moving the ConfigName 'resolver'
into a new 'extract' subpackage only responsible for extracting config names during deployment.