refactor(persistence): Properly split 'bucket' type from config API types
The previous quick and dirty version used the config API persistence type.
The refactoring introduces is as a dedicated type entry.

It still simply uses the same string based unmarshalling as the shorthand API
defintion, but stores a dedicated type.
Further refactoring of the TypeDefinition is possible, but does not seem necessary
to ensure a clear user experience while leaving the code acceptably clean.