fix(convert): Remove manually added escape characters on conversion
As v2 value parameters are fully escaped via JSON marshalling before being placed into
a JSON template, manually added escape chars need to be removed when converting a v1 configuration.

For example if a user had a v1 parameter of \"string\" in a YAML, their expected result for
the JSON payload is "string". They just had to manually escape the slash as well as quote, to
ensure the payload contains an escaped quote character instead of just a quote.

As v2 value parameters aim to simply string handling and auto-escape chars as needed, a value
of string in the YAML will already become "string" in the JSON payload; while a value of
"string" (the result of reading the \"string\" on converting) will become the fully escaped
\"string\" once placed in the JSON payload.
This isn't just not what the user desired for their v1 config, it also very likely results in a
broken payload and 400 API response.

As this feature may have some as of yet unforseen side-effects it is placed behind a feature flag
so it can be toggled OFF if it should needed.